Update environment configuration and testing setup
- Added `APP_ENV` variable to `.env` files to specify runtime mode (local or deployed).
- Updated `pyproject.toml` to adjust test paths and coverage source settings.
- Revised README and documentation to reflect changes in local configuration and testing instructions.
- Removed deprecated `config` module and adjusted related imports in the codebase.
- Enhanced validation in settings to ensure required environment variables are provided for deployed environments.

+++ src/realtime_transcribe_service/shutdown/graceful.py
@@ -14,13 +14,13 @@ log = structlog.get_logger(__name__)
 class GracefulShutdown:
     """Handle SIGTERM/SIGINT, mark drain mode, and notify the main loop to exit."""
 
-    def __init__(self, stop_timeout: int = 120) -> None:
+    def __init__(self, stop_timeout: float = 120.0) -> None:
         self._stop_timeout = stop_timeout
         self._draining = False
         self._shutdown_event = asyncio.Event()
 
     @property
-    def stop_timeout(self) -> int:
+    def stop_timeout(self) -> float:
         """Total graceful shutdown budget in seconds."""
         return self._stop_timeout
